requirements.txt
astronomer-cosmos==0.4.0
soda-core-postgres